Duplicate Code refactor
